# openapi.yaml
openapi: 3.0.3
info:
  title: MeteoAPI
  version: '1.0.0'
servers:
  - url: https://YOUR_API_URL
components:
  schemas:
    Station:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        country:
          type: string
        lat:
          type: number
        lon:
          type: number
    CurrentWeather:
      type: object
      properties:
        time:
          type: string
        temperature_2m:
          type: number
        precipitation:
          type: number
        relative_humidity_2m:
          type: number
        wind_speed_10m:
          type: number
    HistoryWeather:
      type: object
      properties:
        daily:
          type: object
          properties:
            time:
              type: array
              items:
                type: string
            temperature_2m_max:
              type: array
              items:
                type: number
            temperature_2m_min:
              type: array
              items:
                type: number
            precipitation_sum:
              type: array
              items:
                type: number
            wind_speed_10m_max:
              type: array
              items:
                type: number
    ForecastWeather:
      type: object
      properties:
        daily:
          type: object
          properties:
            time:
              type: array
              items:
                type: string
            temperature_2m_max:
              type: array
              items:
                type: number
            temperature_2m_min:
              type: array
              items:
                type: number
            precipitation_sum:
              type: array
              items:
                type: number
            wind_speed_10m_max:
              type: array
              items:
                type: number
paths:
  /current:
    get:
      tags: [Weather]
      summary: Get current weather
      parameters:
        - name: station
          in: query
          schema:
            type: string
        - name: lat
          in: query
          schema:
            type: number
            format: float
        - name: lon
          in: query
          schema:
            type: number
            format: float
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentWeather'
              example:
                time: "2024-06-10T14:00:00Z"
                temperature_2m: 22.5
                precipitation: 0.0
                relative_humidity_2m: 60
                wind_speed_10m: 12.0
        '400':
          description: Missing parameters
        '404':
          description: Station not found
        '502':
          description: Open-Meteo error
  /history:
    get:
      tags: [Weather]
      summary: Get historical weather
      parameters:
        - name: station
          in: query
          schema:
            type: string
        - name: lat
          in: query
          schema:
            type: number
            format: float
        - name: lon
          in: query
          schema:
            type: number
            format: float
        - name: date
          in: query
          schema:
            type: string
            format: date
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryWeather'
              example:
                daily:
                  time: ["2024-06-09"]
                  temperature_2m_max: [22.0]
                  temperature_2m_min: [14.0]
                  precipitation_sum: [0.0]
                  wind_speed_10m_max: [15.0]
        '400':
          description: Missing parameters
        '404':
          description: Station not found
        '502':
          description: Open-Meteo error
  /forecast:
    get:
      tags: [Weather]
      summary: Get weather forecast
      parameters:
        - name: station
          in: query
          schema:
            type: string
        - name: lat
          in: query
          schema:
            type: number
            format: float
        - name: lon
          in: query
          schema:
            type: number
            format: float
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastWeather'
              example:
                daily:
                  time: ["2024-06-10", "2024-06-11"]
                  temperature_2m_max: [24.0, 25.0]
                  temperature_2m_min: [15.0, 16.0]
                  precipitation_sum: [0.0, 0.2]
                  wind_speed_10m_max: [12.0, 14.0]
        '400':
          description: Missing parameters
        '404':
          description: Station not found
        '502':
          description: Open-Meteo error
  /stations:
    get:
      tags: [Stations]
      summary: List all stations
      parameters:
        - name: country
          in: query
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Station'
              example:
                - id: FRPAR
                  name: Paris
                  country: FR
                  lat: 48.8566
                  lon: 2.3522
                - id: USNYC
                  name: New York
                  country: US
                  lat: 40.7128
                  lon: -74.0060
  /station/meta:
    get:
      tags: [Stations]
      summary: Get station metadata
      parameters:
        - name: station
          in: query
          schema:
            type: string
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Station'
              example:
                id: FRPAR
                name: Paris
                country: FR
                lat: 48.8566
                lon: 2.3522
        '404':
          description: Station not found
  /station/nearby:
    get:
      tags: [Stations]
      summary: Get nearby stations
      parameters:
        - name: lat
          in: query
          schema:
            type: number
            format: float
          required: true
        - name: lon
          in: query
          schema:
            type: number
            format: float
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Station'
              example:
                - id: FRPAR
                  name: Paris
                  country: FR
                  lat: 48.8566
                  lon: 2.3522
                - id: FRLYO
                  name: Lyon
                  country: FR
                  lat: 45.75
                  lon: 4.85
  /ping:
    get:
      tags: [Special]
      summary: Healthcheck endpoint (ping)
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  ping:
                    type: string
                    example: pong 